BO6: Removed the api for perks and made it a function
